+++ test/internal/byte_utils_test.dart
@@ -1,4 +1,3 @@
-
 import 'dart:io';
 
 import 'package:kdbx/src/internal/byte_utils.dart';
@@ -12,6 +11,13 @@ void main() {
       writer.writeUint32(1);
       print('result: ' + ByteUtils.toHexList(writer.output.toBytes()));
       expect(writer.output.toBytes(), hasLength(4));
+    });
+    test('uint64', () {
+      final bytes = WriterHelper.singleUint64Bytes(6000);
+      final read = ReaderHelper.singleUint64(bytes);
+      print('read: $read');
+      expect(read, 6000);
+      print('bytes: ${ByteUtils.toHexList(bytes)}');
     });
   });
 }